+++ vector/src/main/java/im/vector/app/core/preference/UserAvatarPreference.kt
@@ -25,14 +25,11 @@ import androidx.preference.PreferenceViewHolder
 import im.vector.app.R
 import im.vector.app.core.extensions.vectorComponent
 import im.vector.app.features.home.AvatarRenderer
-import org.matrix.android.sdk.api.session.Session
-import org.matrix.android.sdk.api.util.MatrixItem
+import org.matrix.android.sdk.api.session.user.model.User
 import org.matrix.android.sdk.api.util.toMatrixItem
 
-open class UserAvatarPreference : Preference {
-
-    internal var mAvatarView: ImageView? = null
-    internal var mSession: Session? = null
+class UserAvatarPreference : Preference {
+    private var mAvatarView: ImageView? = null
     private var mLoadingProgressBar: ProgressBar? = null
 
     private var avatarRenderer: AvatarRenderer = context.vectorComponent().avatarRenderer()
@@ -51,24 +48,11 @@ open class UserAvatarPreference : Preference {
 
     override fun onBindViewHolder(holder: PreferenceViewHolder) {
         super.onBindViewHolder(holder)
-
         mAvatarView = holder.itemView.findViewById(R.id.settings_avatar)
         mLoadingProgressBar = holder.itemView.findViewById(R.id.avatar_update_progress_bar)
-        refreshAvatar()
     }
 
-    open fun refreshAvatar() {
-        val session = mSession ?: return
-        val view = mAvatarView ?: return
-        session.getUser(session.myUserId)?.let {
-            avatarRenderer.render(it.toMatrixItem(), view)
-        } ?: run {
-            avatarRenderer.render(MatrixItem.UserItem(session.myUserId), view)
-        }
-    }
-
-    fun setSession(session: Session) {
-        mSession = session
-        refreshAvatar()
+    fun refreshAvatar(user: User) {
+        mAvatarView?.let { avatarRenderer.render(user.toMatrixItem(), it) }
     }
 }